requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.10.3
python-dotenv==1.0.1
